Fizz Buzz and Index Power from Checkio
Fizz Buzz - "Fizz buzz" is a word game we will use to teach the robots
about division. Let's learn computers.
Index Power - You are given an array with positive numbers and a number
N. You should find the N-th power of the element in the array with the
index N. If N is outside of the array, then return -1. Don't forget that
the first element has the index 0.